Moved discord connection logic inside Account component
